Extend possibilities for YAML parsing and add examples of server config files
The config files (scs_auth.yaml and scs_conf.yaml) describe the locations
of directories with the configuration templates, common variables and
secrets.

All configuration files can use the !scs-relative, !scs-secret and
!scs-common tags to reference files using a relative reference,
reference to the secrets directory and reference to the common
directory respectively. The !scs-expand-env tag is used to
expand environment variables used in a string.

Secrets files only support references to other secrets files, and
allow for the !scs-gen-secret tag, which auto-generates a
32 bytes token, and saves it back to the file. Note that comments
will be lost in this process.