Add a template check during server initialization
This checks if the templates do not contain syntax errors and/or
are compatible with the given env. This prevents unnkowingly
starting a server where some files cannot be accessed because of
configuration errors.

In case specific env file variables do not have defaults, and are
defined in a post request body, you may want to disable this
feature through the config file