Enable (optional) rejection of files with dots in key names
Dots in key names cannot be referenced using the scs syntax, and
are therefore not advised to be used in env-files or configuration
files that are referenced from other files.

This check can opitionally be disabled, since users may want to
use dots in names of sub-keys (e.g. in a list that's under the
key that's actually referenced).