Enable on-demand loading of configuration files
This feature is especially usefull if YAML tag constructors should
be re-evaluated for each invocation (e.g. they get a result from
an external server).

For static files, it's advised to restart the server on changes,
since this will also validate all the env-files, and test the
templates. These validations do not happen at startup of on-demand
loading is enabled.