Improve autonomy of errors and logging module
Previously, error and event ids, related to both the 'auth' and
the 'config' modules were defined in these modules. This changes
removes the dependencies between the two, and allows any
module or blueprint to use the errors and logging modules.

This also improves the naming of variables, functions and classes,
to indicate which are for internal use only.