Optimize config loading view function
This ensures (1) that new templates can be added at runtime, so
the cache: false options are working as expected. Note that even
if cache is enabled, new files can be added, which will then be
cached on first load.

This (2) also ensures that the checking if content should be cached
is now restricted to the init function, where all data is iterated
in case cache is enabled, and to the filecache, which controls
whether to return and store results. This way the view function does
not need to contain logic regarding caching