Add support for using SCS behind reverse proxies
By setting the SCS_REVERSE_PROXY_COUNT environment variable, you
can indicate if the server is running behind one or more
proxies. This is relevant, since behind a proxy, the remote
address needs to be determined from the 'X-Forwarded-For' header.
The count indicates how many values to expect in the header. If
using one proxy, only 1 is expected. This prevents a client from
being able to spoof it's IP by setting a X-Forwarded-For header on
the first request to the reverse proxy.

Note that also the DISABLE_SCS_SSL has been renamed to
renamed to SCS_DISABLE_SSL for consistency.