validate: Add support for wildcard paths; Support escaping wildcard chars
This commit adds support for defining wildcard paths in
scs-validate.yaml. This for example allows you to validate the
parsing of all endpoints ending with .json. The function has also
been extended to support escaping the wildcard character with a
backslash, which also applies to the wildcard-paths defined in
user authorization configurations.