Stop using app.template_dir to load templates for the configs blueprint
This was originally implemented to be able to initialize multiple
apps side-by-side for testing, which meant the template_folder
of the blueprint could not be used since it was already initialized,
and therefore re-used accross apps. This method could possibly
interfere with other blueprints, that can be added as extension to
SCS.

This change implements a Jinja-env specifically for the configs
Blueprint, stored at the app-level.